Move describeError up to the channels subsystem root
It lived under adapters/ because only adapters used it. The manager,
router, persistence and membership-cache need it too, so adapters/ is
the wrong home.

Pure relocation: the helper body is unchanged. The twelve importer
edits are the same one-line specifier update, and they have to ride
along or the move does not typecheck.
