+++ src/channels/describe-error.ts
